Inline Case files for EmptyCodeSpec
